% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tnbc}
\alias{tnbc}
\title{TNBC small dataset. Simulated dataset has been filtered so that 100th most highly variable genes are kept.}
\format{
A data frame with 100 rows and 2500 columns:
}
\source{
Simulated data from Starfysh paper
}
\usage{
tnbc
}
\description{
A small dataset for demonstrating package functionality.
}
\keyword{datasets}
