% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snmf.R
\name{snmf}
\alias{snmf}
\title{Spatially regularized Non-negative Matrix Factorization (sNMF)}
\usage{
snmf(
  counts,
  S,
  k,
  Winit = NULL,
  Hinit = NULL,
  niter = 2000,
  Rupdate_iter = 10,
  tol = 1e-04,
  num_initializations = 10,
  probs = 0.75,
  seed = 42
)
}
\arguments{
\item{counts}{Numeric matrix of counts (features × observations), e.g., genes × spots.}

\item{S}{Numeric spatial similarity matrix, e.g., as returned by \code{load_data}.}

\item{k}{Integer, number of components used for the factorization, e.g. the number of cell-types used for deconvolution.}

\item{Winit}{Initialization of W (default NULL).}

\item{Hinit}{Initialization of H (default NULL).}

\item{niter}{Integer, maximum number of iterations for the NMF algorithm (default 2000).}

\item{Rupdate_iter}{Integer, number of iterations after which dispersion matrix R is updated (default 10).}

\item{tol}{Numeric, convergence tolerance for the NMF algorithm (default 1e-4).}

\item{num_initializations}{Integer, number of random initializations for NMF (default 10).}

\item{probs}{Numeric between 0 and 1, quantile used for normalizing W and H matrices (default 0.75).}

\item{seed}{Integer, random seed for reproducibility (default 42).}
}
\value{
A list with six elements:
\describe{
\item{W}{Numeric matrix (features × k), representing basis vectors for features.}
\item{H}{Numeric matrix (k × observations), representing coefficients for each observation.}
\item{phi}{Numeric matrix (features × observations), representing the inverse-dispersion parameter for each observation $\phi_{ij}$.}
\item{alpha}{Numeric vector of the number of features as dimension, representing the inverse-dispersion contribution of each gene.}
\item{beta}{Numeric vector of the number of observations as dimension, representing the inverse-dispersion contribution of each spot.}
\item{niter}{Number of iterations run until convergence.}
}
}
\description{
Performs non-negative matrix factorization (NMF) on count data, incorporating
spatial relationships between observations through a spatial similarity matrix \code{S}.
}
\details{
This function performs spatially regularized NMF on the input count matrix:
\enumerate{
\item Converts \code{counts} and \code{S} to GPU matrices for efficient computation.
\item Sets the random seed for reproducibility.
\item Runs the \code{SNMF} algorithm on the GPU, with \code{k} factors.
\item Normalizes the W (basis) and H (coefficient) matrices using the specified quantile.
\item Computes a normalized H matrix (\code{HC}) representing proportions per spot.
\item Assigns row and column names to W and H matrices based on the input counts.
}
}
\examples{
\dontrun{
# counts and S returned from load_data()
result <- snmf(counts, S, niter = 1000, Rupdate_iter=20, tol = 1e-4, num_initializations = 5)
W <- result$W
H <- result$H
}

}
